+++ examples/demo-rollup/tests/evm/mod.rs
@@ -1,4 +1,3 @@
-mod balances;
 mod evm_account_abstraction;
 mod evm_balances;
 mod evm_test_helper;